# extract weeklyprofit to extra component and the child compenent is embedded in the dashboard
