+++ Blob/src/app/statistic/statistic-dashboard/statistic-dashboard.component.html
@@ -4,11 +4,8 @@
             Wochenumsatz
         </div>
         <div nz-row>
-            <ngx-charts-area-chart [view]="view" [scheme]="colorScheme" [legend]="legend"
-                [showXAxisLabel]="showXAxisLabel" [showYAxisLabel]="showYAxisLabel" [xAxis]="xAxis" [yAxis]="yAxis"
-                [xAxisLabel]="xAxisLabel" [yAxisLabel]="yAxisLabel" [timeline]="timeline" [results]="weeklyProfits"
-                (select)="onSelect($event)">
-            </ngx-charts-area-chart>
+            <app-statistic-weekly-profits [orders]="orders">
+            </app-statistic-weekly-profits>
         </div>
     </div>
     <div nz-col nzSpan="12">Monatsgewinn</div>
