+++ Blob/src/app/statistic/statistic-dashboard/statistic-dashboard.component.ts
@@ -11,26 +11,7 @@ import { IOrderItem } from 'src/app/interfaces/order/IOrderItem';
   styleUrls: ['./statistic-dashboard.component.less'],
 })
 export class StatisticDashboardComponent implements OnInit {
-  multi: ILineAreaChartItems;
-  weeklyProfits: ILineAreaChartItems[];
-
-  view: any[] = [700, 300];
-
-  // options
-  legend: boolean = false;
-  showLabels: boolean = true;
-  animations: boolean = true;
-  xAxis: boolean = true;
-  yAxis: boolean = true;
-  showYAxisLabel: boolean = true;
-  showXAxisLabel: boolean = false;
-  xAxisLabel: string = 'Wochentag';
-  yAxisLabel: string = 'Umsatz [€]';
-  timeline: boolean = false;
-
-  colorScheme = {
-    domain: ['#5AA454', '#E44D25', '#CFC0BB', '#7aa3e5', '#a8385d', '#aae3f5'],
-  };
+  orders: any;
 
   constructor(private titleService: TitleService, private statisticService: StatisticService, private modalService: NzModalService) {
     this.titleService.Title = 'Statistik';
@@ -44,7 +25,7 @@ export class StatisticDashboardComponent implements OnInit {
     this.statisticService.getAllOrders().subscribe(
       (data) => {
         console.log(data);
-        this.calcWeeklyProfits(data);
+        this.orders = data;
       },
       (error) => {
         console.error(error);
@@ -54,49 +35,5 @@ export class StatisticDashboardComponent implements OnInit {
         });
       }
     );
-  }
-
-  private calcWeeklyProfits(listOfOrders: IOrderItem[]): void {
-    let dateLastWeek = new Date();
-    dateLastWeek.setDate(dateLastWeek.getDate() - 7);
-    dateLastWeek.setHours(0, 0, 0, 0);
-
-    let listOfOrdersThisWeek = listOfOrders.filter((order) => new Date(Date.parse(order.createdAt)) > dateLastWeek);
-
-    console.log(listOfOrdersThisWeek);
-
-    // init profits
-    this.weeklyProfits = [{ name: 'Umsatz', series: [] }];
-    // foreach order add to weeklyProfits
-    listOfOrdersThisWeek.forEach((order) => {
-      let weekday =
-        new Date(Date.parse(order.createdAt)).toLocaleString('de-DE', { weekday: 'long' }) +
-        ` (${new Date(Date.parse(order.createdAt)).toLocaleDateString('de-DE')})`;
-      let profit = order.orderedProducts.reduce((sum, order) => (sum += order.price * order.quantity), 0);
-
-      console.log(this.weeklyProfits[0].series.filter((series) => series.name == weekday));
-
-      // If already a series with the same weekday (eg. monday) only update value, in order to avoid duplicate series with same name.
-      if (this.weeklyProfits[0].series.filter((series) => series.name == weekday).length > 0) {
-        this.weeklyProfits[0].series.forEach((series) => {
-          if (series.name == weekday) {
-            series.value += profit;
-          }
-        });
-      }
-      // Series not created so far.
-      else {
-        this.weeklyProfits[0].series.push({
-          name: weekday,
-          value: profit,
-        });
-      }
-    });
-
-    console.log(this.weeklyProfits);
-  }
-
-  onSelect(event) {
-    console.log(event);
   }
 }
