+++ Blob/src/app/statistic/statistic-weekly-profits/statistic-weekly-profits.component.html
@@ -1,0 +1,4 @@
+<ngx-charts-area-chart [view]="view" [scheme]="colorScheme" [legend]="legend" [showXAxisLabel]="showXAxisLabel"
+    [showYAxisLabel]="showYAxisLabel" [xAxis]="xAxis" [yAxis]="yAxis" [xAxisLabel]="xAxisLabel"
+    [yAxisLabel]="yAxisLabel" [timeline]="timeline" [results]="weeklyProfits">
+</ngx-charts-area-chart>
