+++ Blob/src/app/statistic/statistic.module.ts
@@ -6,9 +6,10 @@ import { StatisticDashboardComponent } from './statistic-dashboard/statistic-das
 
 import { NzGridModule, NzModalModule } from 'ng-zorro-antd';
 import { NgxChartsModule } from '@swimlane/ngx-charts';
+import { StatisticWeeklyProfitsComponent } from './statistic-weekly-profits/statistic-weekly-profits.component';
 
 @NgModule({
-  declarations: [StatisticDashboardComponent],
+  declarations: [StatisticDashboardComponent, StatisticWeeklyProfitsComponent],
   imports: [CommonModule, StatisticRoutingModule, NzGridModule, NgxChartsModule, NzModalModule],
 })
 export class StatisticModule {}
